requirements.txt
google-genai
langchain-google-genai
langchain
langgraph
langsmith
qdrant-client
pypdf
langchain-community
requests
streamlit
tqdm
python-dotenv
pytest
